Module uic_upvl::upvl_db

UPVL Database Interface

The first version of the database contains one table with the provisions.

We assume that the DSK is unique.

We assume that there are so few entries in the database that performance is not a problem. This assumption is based on the fact that the entire list of provisions must be published as one MQTT message.

Structs

UpvlDbEntry

Type that matches the rows in the database provision table.

Constants

CURRENT_DB_VERSION

Current Database Version. At startup db_upgrade checks this against the “pragma_user_version”, and ensure migrating the database to the current version.

LOG_TAG

Functions

db_debug_print_provisions

Print out all the provisions.

db_list_provisions

Print out everything from the data base in json format.

db_remove_entry

Remove a provision by DSK.

db_setup

Open the database connection to the specified file. Return the connection.

db_teardown

Close the database connection.

db_upgrade

Checks the database version and upgrades the data within the database if required

db_upsert_entry

Update database for a provision in deserialized json format. If a provision with this dsk already exists, update the fields. If it does not exist, create a new provison.

entry_merge

Merge two serde_json::value objects. This will overwrite any items in dest that also exist in src, items that doesn’t exist in dest will be created

is_entry_valid

Checks if an entry is valid