Module uic_upvl::upvl_mqtt

We use the mosquitto_client crate for the connection to the broker.

This module contains the main event loop of the UPVL client.

It also provides a bit more wrapping and customization specifically to the UPVL Client features.

Other crates with MQTT client implementations are available, including some implemented in Rust, but their interfaces may be different.

Other crates:

https://github.com/eclipse/paho.mqtt.rust (wrapper for the C client, claims support for v5 from version 0.7.0, SSL)

https://crates.io/crates/actix-mqtt-client (active, at least)

https://docs.rs/crate/rumqtt/0.31.0

https://github.com/inre/rust-mq (Rust client, does not look active)

https://github.com/fluffysquirrels/mqtt-async-client-rs (Rust client, so far 3.1.1, relatively new, MIT License)

Structs

UpvlMqttSession
Type to hold a connection to the broker.

UpvlMqttSessionData
A data-type for the data to send to the mqtt-client Callbacks object. The data is available as the first argument to the callback.

UpvlSubscribers
The subscribers returned by mosquitto_client when we subscribe to a topic. We only use this to access the topic-matching function.

Constants

LOG_TAG
The default upvl_mqtt log tag

PUB_PREFIX
Logging helper

SUBS_PREFIX
Logging helper TODO: fancier logging